Added color constants; Added interfaces folder; Added validation messages; Refactored code;
