+++ src/app/modules/login-module/interfaces/user.ts
@@ -1,4 +1,4 @@
-export class User {
+export interface User {
   login: string;
   password: string;
 }
