+++ src/app/modules/login-module/interfaces/validation-messages.ts
@@ -1,0 +1,4 @@
+export interface ValidationMessages {
+  email: Array<object>,
+  password: Array<object>
+}
