+++ src/app/modules/login-module/login.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from "@angular/common/http";
 import { Observable, Subject } from 'rxjs';
 import { environment } from "../../../environments/environment";
-import { User } from "./user";
+import { User } from "./interfaces/user";
 
 @Injectable({
   providedIn: 'root'
