+++ src/app/modules/login-module/login/login.component.scss
@@ -1,3 +1,5 @@
+@import '../../../../scss/constants/colors';
+
 .main-container {
   width: 100%;
   min-height: 90vh;
@@ -13,8 +15,8 @@ form {
   flex-direction: column;
   flex-wrap: wrap;
   justify-content: center;
-  background-color: #ffffff;
-  border: darkgray 2px solid;
+  background-color: $form_background;
+  border: $form_border 2px solid;
   width: 400px;
   padding: 20px;
 }
@@ -24,7 +26,7 @@ input {
 }
 
 button {
-  background-color: transparent;
+  background-color: $button_background;
   border: none;
   border-radius: 10px;
   font-size: 16px;
@@ -38,10 +40,10 @@ button {
 }
 
 button:hover {
-  box-shadow: 0 12px 16px 0 rgba(0,0,0,0.24), 0 17px 50px 0 rgba(0,0,0,0.19);
+  box-shadow: $button_shadow_hover;
 }
 
 button:active {
-  box-shadow: 0 12px 16px 0 rgba(0,80,100,0.24), 0 17px 50px 0 rgba(0,80,100,0.19);
-  color: rgba(0,0,0,0.24);
+  box-shadow: $button_shadow_active;
+  color: $button_text;
 }
