+++ src/app/modules/share-module/header/header.component.scss
@@ -1,8 +1,10 @@
+@import "../../../../scss/constants/colors";
+
 mat-toolbar {
-  background-color: #fff016;
-  -webkit-box-shadow: 3px 12px 10px -11px rgba(0,0,0,0.84);
-  -moz-box-shadow: 3px 12px 10px -11px rgba(0,0,0,0.84);
-  box-shadow: 3px 12px 10px -11px rgba(0,0,0,0.84);
+  background-color: $header_background;
+  -webkit-box-shadow: $header_box_shadow;
+  -moz-box-shadow: $header_box_shadow;
+  box-shadow: $header_box_shadow;
   display: flex;
   flex-direction: row;
   justify-content: space-between;
