+++ src/scss/constants/colors.scss
@@ -1,0 +1,29 @@
+/*
+*static colors
+*/
+$white: #ffffff;
+$ghostwhite: #f8f8ff;
+$yellow: #fff016;
+$darkgray: #a9a9a9;
+$transparent: transparent;
+
+/*
+*global app component
+*/
+$app_background: $ghostwhite;
+
+/*
+*header component
+*/
+$header_background: $yellow;
+$header_box_shadow: 3px 12px 10px -11px rgba(0,0,0,0.84);
+
+/*
+*login component
+*/
+$form_background: $white;
+$form_border: $darkgray;
+$button_background: $transparent;
+$button_shadow_hover: 0 12px 16px 0 rgba(0,0,0,0.24), 0 17px 50px 0 rgba(0,0,0,0.19);
+$button_shadow_active: 0 12px 16px 0 rgba(0,80,100,0.24), 0 17px 50px 0 rgba(0,80,100,0.19);
+$button_text: rgba(0,0,0,0.24);
