+++ src/styles.scss
@@ -1,11 +1,11 @@
 /* You can add global styles to this file, and also import other style files */
 @import "~@angular/material/prebuilt-themes/indigo-pink.css";
+@import "scss/constants/colors";
 
 html, body { height: 100%; }
 body {
   margin: 0;
   padding: 0;
-  font-family: Roboto,
-  "Helvetica Neue", sans-serif;
-  background-color: ghostwhite;
+  font-family: Roboto, "Helvetica Neue", sans-serif;
+  background-color: $app_background;
 }
